refactor dynamic policies

+++ src/main/kotlin/id/walt/auditor/dynamic/OPAPolicyEngine.kt
@@ -3,21 +3,14 @@ package id.walt.auditor.dynamic
 import com.beust.klaxon.JsonObject
 import com.beust.klaxon.Klaxon
 import com.beust.klaxon.Parser
-import id.walt.common.resolveContentToTempFile
-import io.ktor.client.*
-import io.ktor.client.engine.cio.*
-import io.ktor.client.plugins.contentnegotiation.*
-import io.ktor.client.request.*
-import io.ktor.client.statement.*
-import io.ktor.serialization.kotlinx.json.*
-import kotlinx.coroutines.runBlocking
+import id.walt.common.resolveContentToFile
 import mu.KotlinLogging
 import java.io.File
 
 object OPAPolicyEngine : PolicyEngine {
     private val log = KotlinLogging.logger {}
 
-
+    const val TEMP_PREFIX = "_TEMP_"
 
     fun validate(jsonInput: String, data: Map<String, Any?>, regoPolicy: String, regoQuery: String): Boolean {
         val input: Map<String, Any?> = Parser.default().parse(StringBuilder(jsonInput)) as JsonObject
@@ -26,7 +19,7 @@ object OPAPolicyEngine : PolicyEngine {
     }
 
     override fun validate(input: Map<String, Any?>, data: Map<String, Any?>, policy: String, query: String): Boolean {
-        val regoFile = resolveContentToTempFile(policy)
+        val regoFile = resolveContentToFile(policy, tempPrefix = TEMP_PREFIX, tempPostfix = ".rego")
         val dataFile = File.createTempFile("data", ".json")
         dataFile.writeText(JsonObject(data).toJsonString())
         try {
@@ -38,7 +31,7 @@ object OPAPolicyEngine : PolicyEngine {
             log.debug("rego eval output: {}", output)
             return Klaxon().parseArray<Boolean>(output)?.all { it } ?: false
         } finally {
-            if(regoFile.exists()) {
+            if(regoFile.exists() && regoFile.name.startsWith(TEMP_PREFIX)) {
               regoFile.delete()
             }
             if(dataFile.exists()) {
